requirements.txt
lxml==4.4.1
requests==2.22.0
pytest==5.2.2
pytest-cov==2.8.1
requests-mock==1.7.0
pypandoc==1.4
python-dateutil==2.8.1
